refactor: break down CharacterStatsManager to meet 500-line limit
- Extract CharacterEditingHeader component (110 lines)
- Extract CharacterAbilityScores component (60 lines)
- Extract CharacterNotes component (70 lines)
- Reduce main CharacterStatsManager from 571 to 410 lines
- Fix ESLint issues with unused parameters
- Maintain all existing functionality and test compatibility
- Clean build and lint passing

Addresses Codacy complexity warning while preserving all features.
